assistive-arm/pyproject.toml
[tool.poetry]
name = "assistive-arm"
version = "0.1.0"
description = ""
authors = ["Xabier Irizar <xirizar@seas.harvard.edu>"]
readme = "README.md"
repository = "https://github.com/xabirizar9/assistive-arm"

[tool.poetry.dependencies]
python = "~3.10"
numpy = "^1.25.0"
matplotlib = "^3.7.2"
importlib = "^1.0.4"
pathlib = "^1.0.1"
jupyter = "^1.0.0"
ipykernel = "^6.24.0"
tqdm = "^4.65.0"
sympy = "^1.12"
qtm-rt = "^3.0.1"
pyzmq = "^25.1.0"
gpiozero = "^1.6.2"
pandas = "^2.1.0"
typer = "^0.9.0"
scipy = "^1.11.2"
lxml = "^4.9.3"
ipympl = "^0.9.3"
open3d = "^0.17.0"
scikit-learn = "^1.3.1"
addict = "^2.4.0"
pyquaternion = "^0.9.9"
ipyvolume = "^0.6.3"
bqplot = "^0.12.40"
roboticstoolbox-python = "^1.1.0"
chspy = "^1.2.2"
seaborn = "^0.13.2"
statistics = "^1.0.3.5"
nbstripout = "^0.7.1"
pyarrow = "^15.0.0"
scienceplots = "^2.1.1"
can = "^0.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"